base class restructure
