+++ src/pydeequ/suggestions.py
@@ -4,75 +4,6 @@ import pdb
 from pydeequ.exceptions import JavaClassNotFoundException
 import pydeequ.jvm_conversions as jc
 
-class ConstraintSuggestionRunBuilder:
-    """
-    A class to build a ConstraintSuggestionRun using a fluent API.
-    """
-    def __init__(self, SparkSession, dataFrame):
-        """
-        Args:
-            SparkSession (pyspark.sql.SparkSession)
-            dataFrame (pyspark.sql.dataframe.DataFrame)
-        """
-        self.spark = SparkSession
-        self._dataFrame = dataFrame
-        run_builder = self._jvm.com.amazon.deequ.suggestions.ConstraintSuggestionRunBuilder
-        self.jvmConstraintSuggestionRunBuilder = run_builder(
-            self.dataFrame._jdf
-        )
-
-    @property
-    def _jsparkSession(self):
-        return self.spark._jsparkSession
-
-    @property
-    def _jvm(self):
-        return self.spark.sparkContext._jvm
-
-    @property
-    def dataFrame(self):
-        return self._dataFrame
-
-    def addConstraintRule(self, constraint):
-        """
-        Add a single rule for suggesting constraints based on ColumnProfiles to the run.
-        
-        Args:
-            constraintRule
-        """
-        jvmRule = constraint._jvmRule
-        self.jvmConstraintSuggestionRunBuilder.addConstraintRule(jvmRule())
-        return self
-
-    def run(self):
-        result = self.jvmConstraintSuggestionRunBuilder.run()
-
-        jvmSuggestionResult = self._jvm.com.amazon.deequ \
-            .suggestions.ConstraintSuggestionResult
-        df = jvmSuggestionResult.getConstraintSuggestionsAsJson(
-            result
-        )
-        return df
-
-class ConstraintSuggestionRunner:
-    """
-    """
-    def __init__(self, SparkSession):
-        """
-        Args:
-            SparkSession ():
-        """
-        self.spark = SparkSession
-
-    def onData(self, dataFrame):
-        """
-        Starting point to construct a run on constraint suggestions.
-        
-        Args:
-            dataFrame (pyspark.sql.dataframe.DataFrame):
-            spark dataFrame on which the checks will be verified.
-        """
-        return ConstraintSuggestionRunBuilder(self.spark, dataFrame)
 
 class Rules:
     """
